Change from ALIAS to MATERIALIZED for generated columns, add setting to insert into materialized columns
