+++ sink-connector-lightweight/src/main/java/com/altinity/clickhouse/debezium/embedded/ddl/parser/Constants.java
@@ -2,7 +2,7 @@ package com.altinity.clickhouse.debezium.embedded.ddl.parser;
 
 public class Constants {
 
-    public static final String ALIAS = "ALIAS";
+    public static final String ALIAS = "MATERIALIZED";
     public static final String PARTITION_BY = " PARTITION BY ";
     public static final String ORDER_BY = " ORDER BY ";
     public static final String ORDER_BY_TUPLE = " ORDER BY tuple()";
